requirements.txt
Flask>=1.0.0
google-cloud-datastore==1.7.1
gunicorn==19.9.0
Flask-SQLAlchemy==2.5.0
PyMySQL==0.9.2
Flask-PyMongo>=2.0.0
oauth2client==4.1.2
PyMongo==3.7.2
six==1.13.0
Flask-WTF==0.14.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
